+++ packages/ice/src/service/serverCompiler.ts
@@ -10,7 +10,7 @@ import type { ServerCompiler } from '../types/plugin.js';
 import type { UserConfig } from '../types/userConfig.js';
 import escapeLocalIdent from '../utils/escapeLocalIdent.js';
 import cssModulesPlugin from '../esbuild/cssModules.js';
-import resolvePlugin from '../esbuild/resolve.js';
+import externalPlugin from '../esbuild/external.js';
 import ignorePlugin from '../esbuild/ignore.js';
 import createAssetsPlugin from '../esbuild/assets.js';
 import { CACHE_DIR, SERVER_OUTPUT_DIR } from '../constant.js';
@@ -40,12 +40,23 @@ const { merge, difference } = lodash;
 
 export function createServerCompiler(options: Options) {
   const { task, rootDir, command, server, syntaxFeatures } = options;
-
-  const alias = task.config?.alias || {};
   const externals = task.config?.externals || {};
   const define = task.config?.define || {};
   const sourceMap = task.config?.sourceMap;
   const dev = command === 'start';
+
+  // Filter empty alias.
+  const ignores = [];
+  const taskAlias = task.config?.alias || {};
+  const alias: Record<string, string> = {};
+  Object.keys(taskAlias).forEach((aliasKey) => {
+    const value = taskAlias[aliasKey];
+    if (value) {
+      alias[aliasKey] = value;
+    } else {
+      ignores.push(aliasKey);
+    }
+  });
 
   const defineVars = {};
   // auto stringify define value
@@ -84,6 +95,8 @@ export function createServerCompiler(options: Options) {
     if (preBundle) {
       depsMetadata = await createDepsMetadata({
         task,
+        alias,
+        ignores,
         rootDir,
         // Pass transformPlugins only if syntaxFeatures is enabled
         plugins: enableSyntaxFeatures ? [
@@ -115,6 +128,7 @@ export function createServerCompiler(options: Options) {
       bundle: true,
       format,
       target: 'node12.20.0',
+      alias,
       // enable JSX syntax in .js files by default for compatible with migrate project
       // while it is not recommended
       loader: { '.js': 'jsx' },
@@ -129,8 +143,8 @@ export function createServerCompiler(options: Options) {
       plugins: [
         ...(customBuildOptions.plugins || []),
         emptyCSSPlugin(),
-        resolvePlugin({
-          alias,
+        externalPlugin({
+          ignores,
           externalDependencies: externalDependencies ?? !server.bundle,
           format,
           externals: server.externals,
@@ -206,16 +220,18 @@ interface CreateDepsMetadataOptions {
   rootDir: string;
   task: TaskConfig<Config>;
   plugins: esbuild.Plugin[];
+  alias: Record<string, string>;
+  ignores: string[];
 }
 /**
  *  Create dependencies metadata only when server entry is bundled to esm.
  */
-async function createDepsMetadata({ rootDir, task, plugins }: CreateDepsMetadataOptions) {
+async function createDepsMetadata({ rootDir, task, plugins, alias, ignores }: CreateDepsMetadataOptions) {
   const serverEntry = getServerEntry(rootDir, task.config?.server?.entry);
-  const alias = (task.config?.alias || {}) as TaskConfig<Config>['config']['alias'];
   const deps = await scanImports([serverEntry], {
     rootDir,
     alias,
+    ignores,
     plugins,
   });
 
@@ -237,6 +253,7 @@ async function createDepsMetadata({ rootDir, task, plugins }: CreateDepsMetadata
     cacheDir,
     taskConfig: task.config,
     alias,
+    ignores,
     plugins,
   });