+++ packages/plugin-store/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 1.0.1
+
+- fix: compatible with win32
+
 ## 1.0.0
 
 - feat: init plugin store