+++ packages/webpack-config/CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 1.0.4
 
-- [chore] do not show webpack config on debug mode
+- [fix] format id for win32 compatible
 - [fix] md4 -> md5 while md4 algorithm is not available anymore in NodeJS 17+ 
+- [chore] do not show webpack config on debug mode
 
 ## 1.0.3